Implement the rest of the database repos.
Phew! Looks like we'll be able to start the interesting bits soon :D

Speaking of, if there are any more complex interactions with the repos 
that require multiple steps, then a trait + impl attached to the data 
class structs might be a good idea.

trait: Like an interface
impl: Implementation of a trait (but not always), like a class
struct: sort of a struct, but sort of a class - but can't contain 
methods (see impl). It's weird.