Squash errors about Fluent and multiple threads by cloning
It's not 'thread safe' per say, but by doing these things:

 - Implementing Clone for Translations
 - Using fluent::concurrent::FluentBundle instead

....we've allowed it to work in a 'threadsafe' manner (even if 
'threadsafe' here means "take a copy of it".

It's not a particularly memory-friendly solution, but it'll work for now 
until we gain enough experience to revisit this and improve on our 
approach to GlobalState (which, as it turns out, isn't really 'global' 
at all, since we're cloning it).

Note to self: If we need to keep any mutable global shared state, we'll 
have to think of something else to do here.